Moved utils into index.js
