+++ src/index.js
@@ -1,7 +1,19 @@
 const core = require('@actions/core');
 const github = require('@actions/github');
 const axios = require('axios');
-const { isHumanCommenter, extractUserInfo } = require('./utils');
+function isHumanCommenter(comment) {
+    return (
+        comment &&
+        comment.user &&
+        (comment.user.type === 'User' || comment.user.type === 'Mannequin')
+    );
+}
+
+function extractUserInfo(comment) {
+    const login = comment?.user?.login ?? "unknown";
+    const type = comment?.user?.type ?? "unknown";
+    return { login, type };
+}
 const STALE_PR_THRESHOLD_DAYS = 60;
 const CLOSED_PR_GRACE_PERIOD_MS = 12 * 60 * 60 * 1000;
 const CLOSED_PR_LABEL = 'pr-closed-pending-unassign';
